separate FromFrontend.Types module
